rename directory name from '/sign_up' to '/signup'
